--- errors/errors_suite_test.go
@@ -1,13 +1,0 @@
-package errors_test
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-)
-
-func TestErrors(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Errors Suite")
-}
