+++ kubeinstallutils/crd.go
@@ -1,7 +1,7 @@
 package kubeinstallutils
 
 import (
-	"github.com/solo-io/go-utils/errors"
+	"github.com/rotisserie/eris"
 	"k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1"
 	apiexts "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
@@ -17,7 +17,7 @@ func CrdsFromManifest(crdManifestYaml string) ([]*v1beta1.CustomResourceDefiniti
 	for _, obj := range crdRuntimeObjects {
 		apiExtCrd, ok := obj.(*v1beta1.CustomResourceDefinition)
 		if !ok {
-			return nil, errors.Wrapf(err, "internal error: crd manifest must only contain CustomResourceDefinitions")
+			return nil, eris.Wrap(err, "internal error: crd manifest must only contain CustomResourceDefinitions")
 		}
 		crds = append(crds, apiExtCrd)
 	}
@@ -27,7 +27,7 @@ func CrdsFromManifest(crdManifestYaml string) ([]*v1beta1.CustomResourceDefiniti
 func CreateCrds(apiExts apiexts.Interface, crds ...*v1beta1.CustomResourceDefinition) error {
 	for _, crd := range crds {
 		if _, err := apiExts.ApiextensionsV1beta1().CustomResourceDefinitions().Create(crd); err != nil && !apierrors.IsAlreadyExists(err) {
-			return errors.Wrapf(err, "failed to create crd: %v", crd)
+			return eris.Wrapf(err, "failed to create crd: %v", crd)
 		}
 	}
 	return nil
@@ -37,7 +37,7 @@ func DeleteCrds(apiExts apiexts.Interface, crdNames ...string) error {
 	for _, name := range crdNames {
 		err := apiExts.ApiextensionsV1beta1().CustomResourceDefinitions().Delete(name, &v1.DeleteOptions{})
 		if err != nil && !apierrors.IsNotFound(err) {
-			return errors.Wrapf(err, "failed to delete crd: %v", name)
+			return eris.Wrapf(err, "failed to delete crd: %v", name)
 		}
 	}
 	return nil
