+++ testutils/get_current_file.go
@@ -4,7 +4,7 @@ import (
 	"path/filepath"
 	"runtime"
 
-	"github.com/solo-io/go-utils/errors"
+	"github.com/rotisserie/eris"
 )
 
 // returns the absolute path to the file the caller
@@ -12,7 +12,7 @@ import (
 func GetCurrentFile() (string, error) {
 	_, callerFile, _, ok := runtime.Caller(1)
 	if !ok {
-		return "", errors.Errorf("failed to get runtime.Caller")
+		return "", eris.New("failed to get runtime.Caller")
 	}
 	return filepath.Abs(callerFile)
 }
